fix cast warning on 64-bit
git-svn-id: http://svn.eol.ucar.edu/svn/rdss/trunk/zebra/source@8318 3f9b6c58-9310-455a-aec6-8b30269ab333

+++ src/realtime/Archiver/Archiver.cc
@@ -2067,7 +2067,7 @@ SuspendWrite(int waitsecs)
 void
 SuspendWrite (Widget w, XtPointer call_data)
 {
-	int waitsecs = (int) call_data;
+	long waitsecs = (long) call_data;
 	ZebTime now;
 	static ZebTime ends;		/* Time this suspension ends */
 					/* Must be static since the
